close #81 [modify]: Redefine domains.
