+++ src/domain/address.go
@@ -1,11 +1,10 @@
 package domain
 
-type DBAddress struct {
-	ID         int
-	Area       int // 大まかな地域，四国や東北等
-	Prefecture int
-	City       int
-	Name       string // 名前
+type Place struct {
+	ID         int    `json:"id,omitempty"`
+	Area       string `json:"area"`
+	Prefecture string `json:"prefecture"`
+	City       string `json:"city"`
 }
 
 type Address struct {
@@ -14,8 +13,4 @@ type Address struct {
 	Prefecture  string `json:"prefecture"`
 	City        string `json:"city"`
 	Description string `json:"description"`
-}
-
-func (DBAddress) TableName() string {
-	return "ADDRESS"
 }
